--- a/packages/es-dev-server/demo/performance/module-a.js
+++ b/packages/es-dev-server/demo/performance/module-a.js
@@ -1,0 +1,10 @@
+/* eslint-disable */
+import * as _ from 'lodash-es';
+import * as litElement from 'lit-element';
+import * as sinon from 'sinon';
+import * as chai from '@bundled-es-modules/chai';
+
+console.log(_);
+console.log(litElement);
+console.log(sinon);
+console.log(chai);