--- a/packages/es-dev-server/src/create-middlewares.js
+++ b/packages/es-dev-server/src/create-middlewares.js
@@ -2,6 +2,7 @@ import koaStatic from 'koa-static';
 import koaEtag from 'koa-etag';
 import { createBasePathMiddleware } from './middleware/base-path.js';
 import { createHistoryAPIFallbackMiddleware } from './middleware/history-api-fallback.js';
+import { createNodeResolveMiddleware } from './middleware/node-resolve.js';
 import { createBabelMiddleware } from './middleware/babel.js';
 import { createReloadBrowserMiddleware } from './middleware/reload-browser.js';
 import { createTransformIndexHTMLMiddleware } from './middleware/transform-index-html.js';
@@ -26,6 +27,7 @@ export function createMiddlewares(config, fileWatcher) {
     basePath,
     moduleDirectories,
     nodeResolve,
+    preserveSymlinks,
     readUserBabelConfig,
     customBabelConfig,
     watch,
@@ -36,7 +38,7 @@ export function createMiddlewares(config, fileWatcher) {
     watchExcludes,
     watchDebounce,
     customMiddlewares,
-    logBabelErrors,
+    logCompileErrors,
   } = config;
 
   /** @type {import('koa').Middleware[]} */
@@ -56,13 +58,12 @@ export function createMiddlewares(config, fileWatcher) {
 
   const setupBabel =
     customBabelConfig ||
-    nodeResolve ||
     [compatibilityModes.ALL, compatibilityModes.MODERN].includes(compatibilityMode) ||
     readUserBabelConfig;
   const setupCompatibility = compatibilityMode && compatibilityMode !== compatibilityModes.NONE;
-  const setupTransformIndexHTML = setupBabel || setupCompatibility;
+  const setupTransformIndexHTML = nodeResolve || setupBabel || setupCompatibility;
   const setupHistoryFallback = appIndex;
-  const setupMessageChanel = watch || setupBabel;
+  const setupMessageChanel = nodeResolve || watch || setupBabel;
 
   // strip application base path from requests
   if (config.basePath) {
@@ -104,14 +105,27 @@ export function createMiddlewares(config, fileWatcher) {
       createBabelMiddleware({
         rootDir,
         moduleDirectories,
-        nodeResolve,
         readUserBabelConfig,
         compatibilityMode,
         extraFileExtensions,
         customBabelConfig,
         babelExclude,
         babelModernExclude,
-        logBabelErrors,
+        logCompileErrors,
+      }),
+    );
+  }
+
+  if (nodeResolve) {
+    middlewares.push(
+      createNodeResolveMiddleware({
+        rootDir,
+        extraFileExtensions,
+        moduleDirectories,
+        preserveSymlinks,
+        // babel will cache as well, no need to do it here
+        cacheCompilation: !setupBabel,
+        logCompileErrors,
       }),
     );
   }