--- a/packages/es-dev-server/src/middleware/babel.js
+++ b/packages/es-dev-server/src/middleware/babel.js
@@ -1,7 +1,6 @@
 import minimatch from 'minimatch';
-import path from 'path';
 import stripAnsi from 'strip-ansi';
-import { getBodyAsString, toFilePath } from '../utils/utils.js';
+import { getBodyAsString, getRequestFilePath } from '../utils/utils.js';
 import { compatibilityModes, baseFileExtensions } from '../constants.js';
 import { sendMessageToActiveBrowsers } from './message-channel.js';
 import createBabelCompiler from '../utils/babel-compiler.js';
@@ -10,14 +9,13 @@ import createBabelCompiler from '../utils/babel-compiler.js';
  * @typedef {object} BabelMiddlewareConfig
  * @property {string} rootDir
  * @property {string[]} moduleDirectories
- * @property {boolean} nodeResolve
  * @property {boolean} readUserBabelConfig
  * @property {string} compatibilityMode
  * @property {object} [customBabelConfig]
  * @property {string[]} extraFileExtensions
  * @property {string[]} babelModernExclude patterns to exclude from modern babel compilation
  * @property {string[]} babelExclude patterns to exclude from all babel compilation
- * @property {boolean} logBabelErrors
+ * @property {boolean} logCompileErrors
  */
 
 /**
@@ -32,21 +30,6 @@ function shouldCompile(file, legacy, cfg) {
   }
 
   return legacy || !cfg.babelModernExclude.some(pattern => minimatch(file, pattern));
-}
-
-/**
- * @param {import('koa').Context} ctx
- * @param {BabelMiddlewareConfig} cfg
- * @returns {string}
- */
-function getFilePath(ctx, cfg) {
-  if (ctx.url.includes('/inline-module-') && ctx.url.includes('?source=')) {
-    const url = ctx.url.split('?')[0];
-    const indexPath = toFilePath(url);
-    return path.join(cfg.rootDir, indexPath);
-  }
-
-  return ctx.body && ctx.body.path;
 }
 
 /**
@@ -76,7 +59,6 @@ export function createBabelMiddleware(cfg) {
     if (baseURL.startsWith('/polyfills') || !fileExtensions.some(ext => baseURL.endsWith(ext))) {
       return next();
     }
-
     await next();
 
     // should be a 2xx response
@@ -84,7 +66,7 @@ export function createBabelMiddleware(cfg) {
       return undefined;
     }
 
-    const filePath = getFilePath(ctx, cfg);
+    const filePath = getRequestFilePath(ctx, cfg.rootDir);
     const lastModified = ctx.response.headers['last-modified'];
     if (!filePath || !lastModified) {
       return undefined;
@@ -132,9 +114,9 @@ export function createBabelMiddleware(cfg) {
       // strip babel ansi color codes because they're not colored correctly for the browser terminal
       sendMessageToActiveBrowsers('error-message', JSON.stringify(stripAnsi(errorMessage)));
 
-      if (cfg.logBabelErrors) {
+      if (cfg.logCompileErrors) {
         /* eslint-disable-next-line no-console */
-        console.error(errorMessage);
+        console.error(error);
       }
     }
     return undefined;