--- a/packages/es-dev-server/test/snapshots/resolve-module-imports/folder-index-js.js
+++ b/packages/es-dev-server/test/snapshots/resolve-module-imports/folder-index-js.js
@@ -1,0 +1,3 @@
+
+      import '../node_modules/my-module/bar/index.js';
+