… packages/building-utils/forked-babel-plugin-bare-import-rewrite.js
… packages/es-dev-server/README.md
… packages/es-dev-server/demo/performance/index.html
… packages/es-dev-server/package.json
… packages/es-dev-server/src/utils/es-module-lexer.js
… packages/es-dev-server/test/middleware/babel.test.js
… packages/es-dev-server/test/middleware/node-resolve.test.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/basic-exports.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/basic-imports.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/comments.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/dynamic imports.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/file-extension.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/import-in-code.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/import-meta.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/module-field.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/non-bare.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/preserve-symlink.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/resolve-symlink.js
… packages/es-dev-server/test/snapshots/resolve-module-imports/urls.js
… packages/es-dev-server/test/snapshots/transform-index-html/import-map-all.html
… packages/es-dev-server/test/snapshots/transform-index-html/import-map-modern.html
… packages/es-dev-server/test/snapshots/transform-index-html/inline-module-all.html
… packages/es-dev-server/test/snapshots/transform-index-html/inline-module-modern.html
… packages/es-dev-server/test/snapshots/transform-index-html/inline-module-none-babel.html
… packages/es-dev-server/test/snapshots/transform-index-html/simple-all.html
… packages/es-dev-server/test/snapshots/transform-index-html/simple-modern.html
… packages/es-dev-server/test/utils/resolve-module-imports.test.js
… yarn.lock