The current implementation of node-resolve uses babel. This is wasteful, as we're creating a full AST of each js file just for resolving the imports.

In this implementation we use the lexer from [es-module-shims](https://github.com/guybedford/es-module-shims/blob/master/src/lexer.js). This just extracts information about where imports and dynamic imports are, so that we can resolve them and do a simple string replacement. 

There is discussion on separating the lexer https://github.com/guybedford/es-module-shims/issues/35 but that may take a while, so I included it here until we resolve that.

Depending on the project, this can significantly speed up performance. Running the dev server and importing all of `lodash-es` takes 5200ms to load on master, and it takes 1800ms in this branch. Running large bundled libraries like sinon, chai, axe etc. through babel take much longer so that's a huge boost there.

On my machine, running the `open-wc` tests on master takes 5560ms while on this branch it takes 3670ms. This is without coverage turned on, because that still boots up babel. I also updated `karma-esm` to only run babel on files that need coverage. That brings it down to 4400ms with coverage enabled.

Fixes https://github.com/open-wc/open-wc/issues/664