@LarsDenBakker 2019-08-11
We can remove our local copy when https://github.com/guybedford/es-module-lexer/pull/1 is merged